pi_edge_node/requirements.txt
gpiozero==2.0.1
paho-mqtt>=2.0.0